+++ packages/editor/src/components/more-menu/index.js
@@ -1,21 +1,32 @@
 import { __, _x } from '@wordpress/i18n';
 import { useSelect, useDispatch } from '@wordpress/data';
-import { displayShortcut } from '@wordpress/keycodes';
-import { external, moreVertical } from '@wordpress/icons';
-import { MenuGroup, MenuItem, DropdownMenu } from '@wordpress/components';
-import {
-	PreferenceToggleMenuItem,
-	store as preferencesStore,
-} from '@wordpress/preferences';
+import { useState } from '@wordpress/element';
+import { Button } from '@wordpress/components';
+import { moreVertical } from '@wordpress/icons';
+import { store as preferencesStore } from '@wordpress/preferences';
 import { store as interfaceStore, ActionItem } from '@wordpress/interface';
-import { VisuallyHidden } from '@wordpress/ui';
+// eslint-disable-next-line @wordpress/use-recommended-components
+import { Menu } from '@wordpress/ui';
 import CopyContentMenuItem from './copy-content-menu-item';
 import ModeSwitcher from '../mode-switcher';
+import PluginsMenuGroup from './plugins-menu-group';
+import MoreMenuPreferenceItem from './more-menu-preference-item';
 import ToolsMoreMenuGroup from './tools-more-menu-group';
 import ViewMoreMenuGroup from './view-more-menu-group';
 import { store as editorStore } from '../../store';
+import { getKeyboardShortcut } from '../../utils/keyboard-shortcut';
+
+const DISTRACTION_FREE_SHORTCUT = getKeyboardShortcut( {
+	character: '\\',
+	modifier: 'primaryShift',
+} );
+const KEYBOARD_SHORTCUTS_SHORTCUT = getKeyboardShortcut( {
+	character: 'h',
+	modifier: 'access',
+} );
 
 export default function MoreMenu( { isRevisionMode = false } ) {
+	const [ isOpen, setIsOpen ] = useState( false );
 	const { openModal } = useDispatch( interfaceStore );
 	const { set: setPreference } = useDispatch( preferencesStore );
 	const { toggleDistractionFree } = useDispatch( editorStore );
@@ -25,136 +36,125 @@ export default function MoreMenu( { isRevisionMode = false } ) {
 		[]
 	);
 
+	const onClose = () => setIsOpen( false );
 	const turnOffDistractionFree = () => {
 		setPreference( 'core', 'distractionFree', false );
 	};
-	const dropdownProps = {
-		icon: moreVertical,
-		label: __( 'Options' ),
-		popoverProps: {
-			placement: 'bottom-end',
-			className: 'more-menu-dropdown__content',
-		},
-		toggleProps: {
-			showTooltip: ! showIconLabels,
-			...( showIconLabels && { variant: 'tertiary' } ),
-			tooltipPosition: 'bottom',
-			size: 'compact',
-		},
-	};
+
+	const trigger = (
+		<Menu.Trigger
+			render={
+				<Button
+					size="compact"
+					icon={ moreVertical }
+					label={ __( 'Options' ) }
+					showTooltip={ ! showIconLabels }
+					tooltipPosition="bottom"
+					variant={ showIconLabels ? 'tertiary' : undefined }
+				/>
+			}
+		/>
+	);
+	const positioner = <Menu.Positioner align="end" />;
 
 	if ( isRevisionMode ) {
 		return (
-			<DropdownMenu { ...dropdownProps }>
-				{ () => <ModeSwitcher /> }
-			</DropdownMenu>
+			<Menu.Root open={ isOpen } onOpenChange={ setIsOpen }>
+				{ trigger }
+				<Menu.Popup
+					className="editor-more-menu__popup"
+					positioner={ positioner }
+				>
+					<ModeSwitcher />
+				</Menu.Popup>
+			</Menu.Root>
 		);
 	}
 
 	return (
-		<>
-			<DropdownMenu { ...dropdownProps }>
-				{ ( { onClose } ) => (
-					<>
-						<MenuGroup label={ _x( 'View', 'noun' ) }>
-							<PreferenceToggleMenuItem
-								scope="core"
-								name="fixedToolbar"
-								onToggle={ turnOffDistractionFree }
-								label={ __( 'Top toolbar' ) }
-								info={ __(
-									'Access all block and document tools in a single place'
-								) }
-								messageActivated={ __(
-									'Top toolbar activated.'
-								) }
-								messageDeactivated={ __(
-									'Top toolbar deactivated.'
-								) }
-							/>
-							<PreferenceToggleMenuItem
-								scope="core"
-								name="distractionFree"
-								label={ __( 'Distraction free' ) }
-								info={ __( 'Write with calmness' ) }
-								handleToggling={ false }
-								onToggle={ () =>
-									toggleDistractionFree( {
-										createNotice: false,
-									} )
-								}
-								messageActivated={ __(
-									'Distraction free mode activated.'
-								) }
-								messageDeactivated={ __(
-									'Distraction free mode deactivated.'
-								) }
-								shortcut={ displayShortcut.primaryShift(
-									'\\'
-								) }
-							/>
-							<PreferenceToggleMenuItem
-								scope="core"
-								name="focusMode"
-								label={ __( 'Spotlight mode' ) }
-								info={ __( 'Focus on one block at a time' ) }
-								messageActivated={ __(
-									'Spotlight mode activated.'
-								) }
-								messageDeactivated={ __(
-									'Spotlight mode deactivated.'
-								) }
-							/>
-							<ViewMoreMenuGroup.Slot fillProps={ { onClose } } />
-						</MenuGroup>
-						<ModeSwitcher />
-						<ActionItem.Slot
-							name="core/plugin-more-menu"
-							label={ __( 'Panels' ) }
-							fillProps={ { onClick: onClose } }
-						/>
-						<MenuGroup label={ __( 'Tools' ) }>
-							<MenuItem
-								onClick={ () =>
-									openModal( 'editor/keyboard-shortcut-help' )
-								}
-								shortcut={ displayShortcut.access( 'h' ) }
-							>
-								{ __( 'Keyboard shortcuts' ) }
-							</MenuItem>
-							<CopyContentMenuItem />
-							<MenuItem
-								icon={ external }
-								href={ __(
-									'https://wordpress.org/documentation/article/wordpress-block-editor/'
-								) }
-								target="_blank"
-								rel="noopener"
-							>
-								{ __( 'Help' ) }
-								<VisuallyHidden render={ <span /> }>
-									{
-										/* translators: accessibility text */
-										__( '(opens in a new tab)' )
-									}
-								</VisuallyHidden>
-							</MenuItem>
-							<ToolsMoreMenuGroup.Slot
-								fillProps={ { onClose } }
-							/>
-						</MenuGroup>
-						<MenuGroup>
-							<MenuItem
-								onClick={ () =>
-									openModal( 'editor/preferences' )
-								}
-							>
-								{ __( 'Preferences' ) }
-							</MenuItem>
-						</MenuGroup>
-					</>
-				) }
-			</DropdownMenu>
-		</>
+		<Menu.Root open={ isOpen } onOpenChange={ setIsOpen }>
+			{ trigger }
+			<Menu.Popup
+				className="editor-more-menu__popup"
+				positioner={ positioner }
+			>
+				<Menu.Group>
+					<Menu.GroupLabel>{ _x( 'View', 'noun' ) }</Menu.GroupLabel>
+					<MoreMenuPreferenceItem
+						scope="core"
+						name="fixedToolbar"
+						onToggle={ turnOffDistractionFree }
+						label={ __( 'Top toolbar' ) }
+						info={ __(
+							'Access all block and document tools in a single place'
+						) }
+						messageActivated={ __( 'Top toolbar activated.' ) }
+						messageDeactivated={ __( 'Top toolbar deactivated.' ) }
+					/>
+					<MoreMenuPreferenceItem
+						scope="core"
+						name="distractionFree"
+						label={ __( 'Distraction free' ) }
+						info={ __( 'Write with calmness' ) }
+						handleToggling={ false }
+						onToggle={ () =>
+							toggleDistractionFree( { createNotice: false } )
+						}
+						messageActivated={ __(
+							'Distraction free mode activated.'
+						) }
+						messageDeactivated={ __(
+							'Distraction free mode deactivated.'
+						) }
+						shortcut={ DISTRACTION_FREE_SHORTCUT }
+					/>
+					<MoreMenuPreferenceItem
+						scope="core"
+						name="focusMode"
+						label={ __( 'Spotlight mode' ) }
+						info={ __( 'Focus on one block at a time' ) }
+						messageActivated={ __( 'Spotlight mode activated.' ) }
+						messageDeactivated={ __(
+							'Spotlight mode deactivated.'
+						) }
+					/>
+					<ViewMoreMenuGroup.Slot fillProps={ { onClose } } />
+				</Menu.Group>
+				<Menu.Separator />
+				<ModeSwitcher />
+				<ActionItem.Slot
+					name="core/plugin-more-menu"
+					label={ __( 'Panels' ) }
+					as={ PluginsMenuGroup }
+					fillProps={ { onClick: onClose } }
+				/>
+				<Menu.Separator />
+				<Menu.Group>
+					<Menu.GroupLabel>{ __( 'Tools' ) }</Menu.GroupLabel>
+					<Menu.Item
+						onClick={ () =>
+							openModal( 'editor/keyboard-shortcut-help' )
+						}
+						shortcut={ KEYBOARD_SHORTCUTS_SHORTCUT }
+					>
+						{ __( 'Keyboard shortcuts' ) }
+					</Menu.Item>
+					<CopyContentMenuItem />
+					<Menu.LinkItem
+						href={ __(
+							'https://wordpress.org/documentation/article/wordpress-block-editor/'
+						) }
+						openInNewTab
+					>
+						{ __( 'Help' ) }
+					</Menu.LinkItem>
+					<ToolsMoreMenuGroup.Slot fillProps={ { onClose } } />
+				</Menu.Group>
+				<Menu.Separator />
+				<Menu.Item onClick={ () => openModal( 'editor/preferences' ) }>
+					{ __( 'Preferences' ) }
+				</Menu.Item>
+			</Menu.Popup>
+		</Menu.Root>
 	);
 }