+++ packages/editor/src/private-apis.js
@@ -12,6 +12,8 @@ import PostCardPanel from './components/post-card-panel';
 import PreferencesModal from './components/preferences-modal';
 import { usePostActions } from './components/post-actions/actions';
 import usePostFields from './components/post-fields';
+import MoreMenuItem from './components/more-menu/more-menu-item';
+import MoreMenuPreferenceItem from './components/more-menu/more-menu-preference-item';
 import ToolsMoreMenuGroup from './components/more-menu/tools-more-menu-group';
 import ViewMoreMenuGroup from './components/more-menu/view-more-menu-group';
 import ResizableEditor from './components/resizable-editor';
@@ -36,6 +38,8 @@ lock( privateApis, {
 	PreferencesModal,
 	usePostActions,
 	usePostFields,
+	MoreMenuItem,
+	MoreMenuPreferenceItem,
 	ToolsMoreMenuGroup,
 	ViewMoreMenuGroup,
 	ResizableEditor,